+++ jkube-kit/common/src/main/java/org/eclipse/jkube/kit/common/util/YamlUtil.java
@@ -16,9 +16,6 @@ package org.eclipse.jkube.kit.common.util;
 import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -30,6 +27,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import org.apache.commons.lang3.StringUtils;
 
+import static org.eclipse.jkube.kit.common.util.MapUtil.getFlattenedMap;
+
 public class YamlUtil {
 
   private static final YAMLFactory YAML_FACTORY = new YAMLFactory();
@@ -39,9 +38,9 @@ public class YamlUtil {
   private YamlUtil() {
   }
 
-    protected static Properties getPropertiesFromYamlResource(URL resource) {
-        return getPropertiesFromYamlResource(null, resource);
-    }
+  public static Properties getPropertiesFromYamlResource(URL resource) {
+      return getPropertiesFromYamlResource(null, resource);
+  }
 
   protected static Properties getPropertiesFromYamlResource(String activeProfile, URL resource) {
     if (resource != null) {
@@ -58,58 +57,6 @@ public class YamlUtil {
     }
     return new Properties();
   }
-
-    /**
-     * Build a flattened representation of the Yaml tree. The conversion is compliant with the thorntail spring-boot rules.
-     */
-    private static Map<String, Object> getFlattenedMap(Map<?, ?> source) {
-        Map<String, Object> result = new LinkedHashMap<>();
-        buildFlattenedMap(result, source, null);
-        return result;
-    }
-
-    private static void buildFlattenedMap(Map<String, Object> result, Map<?, ?> source, String path) {
-        for (Map.Entry<?, ?> entry : source.entrySet()) {
-            Object keyObject = entry.getKey();
-
-            String key;
-            if (keyObject instanceof String) {
-                key = (String) keyObject;
-            } else if (keyObject instanceof Number) {
-                key = String.valueOf(keyObject);
-            } else {
-                // If user creates a wrong application.yml then we get a runtime classcastexception
-                throw new IllegalArgumentException(String.format("Expected to find a key of type String but %s with content %s found.",
-                        keyObject.getClass(), keyObject.toString()));
-            }
-
-            if (path !=null && path.trim().length()>0) {
-                if (key.startsWith("[")) {
-                    key = path + key;
-                }
-                else {
-                    key = path + "." + key;
-                }
-            }
-            Object value = entry.getValue();
-            if (value instanceof Map) {
-
-                Map<?, ?> map = (Map<?, ?>) value;
-                buildFlattenedMap(result, map, key);
-            }
-            else if (value instanceof Collection) {
-                Collection<?> collection = (Collection<?>) value;
-                int count = 0;
-                for (Object object : collection) {
-                    buildFlattenedMap(result,
-                            Collections.singletonMap("[" + (count++) + "]", object), key);
-                }
-            }
-            else {
-                result.put(key, (value != null ? value.toString() : ""));
-            }
-        }
-    }
 
   public static Properties getPropertiesFromYamlString(String yamlString) throws IOException {
     final Properties properties = new Properties();